+++ probium/engines/zip_office.py
@@ -1,7 +1,7 @@
 #made for zip files - recursive scan example engine
 from __future__ import annotations
 import zipfile, io
-from ..types import Candidate, Result
+from ..models import Candidate, Result
 from .base import EngineBase
 from ..registry import register
